sql instancestore

+++ Daenet.DurableTask.SqlStateProvider/WorkItemTable.sql
@@ -3,7 +3,7 @@
 	[Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [InstanceId] NVARCHAR(50) NOT NULL, 
     [ExeutionId] NVARCHAR(50) NOT NULL, 
-    [SequenceNumber] INT NOT NULL, 
+    [SequenceNumber] BIGINT NOT NULL, 
     [HistoryEvent] NVARCHAR(MAX) NOT NULL, 
     [EventTimestamp] DATETIME NOT NULL,
 	[TimeStamp] DATETIMEOFFSET NOT NULL
